+++ crates/re_space_view/Cargo.toml
@@ -22,18 +22,13 @@ all-features = true
 default = []
 
 [dependencies]
-re_log.workspace = true
 re_log_types.workspace = true
 re_data_store.workspace = true
 re_entity_db.workspace = true
-re_query.workspace = true
 re_tracing.workspace = true
 re_types_core.workspace = true
-re_types.workspace = true
+re_ui.workspace = true
 re_viewer_context.workspace = true
+re_viewport_blueprint.workspace = true
 
 egui.workspace = true
-itertools.workspace = true
-nohash-hasher.workspace = true
-slotmap.workspace = true
-smallvec = { workspace = true, features = ["serde"] }
