+++ crates/re_space_view/src/lib.rs
@@ -3,23 +3,15 @@
 //! Types & utilities for defining Space View classes and communicating with the Viewport.
 
 pub mod controls;
-mod data_query;
+
 mod heuristics;
 mod screenshot;
-mod space_view;
-mod space_view_contents;
-mod view_properties; // TODO(andreas): better name before `sub_archetype` sticks around?
+mod view_property_ui;
 mod visualizable;
 
-pub use data_query::{DataQuery, EntityOverrideContext, PropertyResolver};
 pub use heuristics::suggest_space_view_for_each_entity;
 pub use screenshot::ScreenshotMode;
-pub use space_view::SpaceViewBlueprint;
-pub use space_view_contents::SpaceViewContents;
-pub use view_properties::{
-    edit_blueprint_component, entity_path_for_view_property, get_blueprint_component,
-    query_view_property, query_view_property_or_default, view_property, view_property_ui,
-};
+pub use view_property_ui::view_property_ui;
 pub use visualizable::determine_visualizable_entities;
 
 pub mod external {
