+++ crates/re_space_view_spatial/Cargo.toml
@@ -36,6 +36,7 @@ re_types = { workspace = true, features = ["ecolor", "glam", "image"] }
 re_tracing.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
+re_viewport_blueprint.workspace = true
 re_space_view.workspace = true
 
 ahash.workspace = true
