+++ crates/re_space_view_time_series/Cargo.toml
@@ -30,6 +30,7 @@ re_tracing.workspace = true
 re_types = { workspace = true, features = ["egui_plot"] }
 re_ui.workspace = true
 re_viewer_context.workspace = true
+re_viewport_blueprint.workspace = true
 
 egui.workspace = true
 egui_plot.workspace = true
