+++ crates/re_space_view_time_series/src/space_view_class.rs
@@ -5,7 +5,7 @@ use egui_plot::{Legend, Line, Plot, PlotPoint, Points};
 use re_data_store::TimeType;
 use re_format::next_grid_tick_magnitude_ns;
 use re_log_types::{EntityPath, TimeInt, TimeZone};
-use re_space_view::{controls, query_view_property_or_default, view_property_ui};
+use re_space_view::{controls, view_property_ui};
 use re_types::blueprint::archetypes::PlotLegend;
 use re_types::{components::Range1D, datatypes::TimeRange, SpaceViewClassIdentifier, View};
 use re_ui::list_item;
@@ -19,6 +19,7 @@ use re_viewer_context::{
     SpaceViewSystemExecutionError, SystemExecutionOutput, ViewQuery, ViewSystemIdentifier,
     ViewerContext, VisualizableEntities,
 };
+use re_viewport_blueprint::query_view_property_or_default;
 
 use crate::line_visualizer_system::SeriesLineSystem;
 use crate::point_visualizer_system::SeriesPointSystem;
