+++ crates/re_viewer/Cargo.toml
@@ -74,6 +74,7 @@ re_types_core.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
 re_viewport.workspace = true
+re_viewport_blueprint.workspace = true
 re_ws_comms = { workspace = true, features = ["client"] }
 
 # Internal (optional):
