+++ crates/re_viewer/src/ui/selection_panel.rs
@@ -268,7 +268,7 @@ fn data_section_ui(item: &Item) -> Option<Box<dyn DataUi>> {
 fn space_view_button(
     ctx: &ViewerContext<'_>,
     ui: &mut egui::Ui,
-    space_view: &re_space_view::SpaceViewBlueprint,
+    space_view: &re_viewport_blueprint::SpaceViewBlueprint,
 ) -> egui::Response {
     let item = Item::SpaceView(space_view.id);
     let is_selected = ctx.selection().contains_item(&item);
