+++ crates/re_viewport_blueprint/Cargo.toml
@@ -4,7 +4,7 @@ description = "The central viewport panel of the Rerun viewer."
 edition.workspace = true
 homepage.workspace = true
 license.workspace = true
-name = "re_viewport"
+name = "re_viewport_blueprint"
 publish = true
 readme = "README.md"
 repository.workspace = true
@@ -20,34 +20,16 @@ all-features = true
 
 [dependencies]
 re_data_store.workspace = true
-re_data_ui.workspace = true
 re_entity_db = { workspace = true, features = ["serde"] }
 re_log_types.workspace = true
 re_log.workspace = true
-re_renderer = { workspace = true, default-features = false, features = [
-  "arrow",
-  "import-gltf",
-  "import-obj",
-  "serde",
-] }
-re_smart_channel.workspace = true
-re_space_view.workspace = true
 re_tracing.workspace = true
 re_types.workspace = true
-re_types_blueprint.workspace = true
 re_types_core.workspace = true
-re_ui.workspace = true
 re_viewer_context.workspace = true
-re_viewport_blueprint.workspace = true
 
-ahash.workspace = true
-egui_tiles.workspace = true
 egui.workspace = true
-glam.workspace = true
-image = { workspace = true, default-features = false, features = ["png"] }
 itertools.workspace = true
 nohash-hasher.workspace = true
-once_cell.workspace = true
-rayon.workspace = true
+slotmap.workspace = true
 smallvec.workspace = true
-static_assertions.workspace = true
