+++ crates/re_viewport_blueprint/src/space_view_contents.rs
